requirements.txt
yt-dlp>=2024.1.0
ffmpeg-python>=0.2.0
openai>=1.0.0
pydub>=0.25.1
click>=8.1.0
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0